requirements.txt
fastapi[all]
Pillow
python-multipart
